requirements.txt
flask
psutil
pytz
pyodbc
python-dotenv
requests
google-generativeai
